refactor(auth): extract session bootstrap
